+++ src/exchanges/cn/bybit/BybitWebSocketClient.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using CCXT.Collector.Core.Abstractions;
+using CCXT.Collector.Models.WebSocket;
 using CCXT.Collector.Library;
 using CCXT.Collector.Service;
 using System.Text.Json;