+++ src/exchanges/cn/hitbtc/HitbtcWebSocketClient.cs
@@ -6,6 +6,7 @@ using CCXT.Collector.Core.Abstractions;
 using CCXT.Collector.Library;
 using CCXT.Collector.Service;
 using System.Text.Json;
+using CCXT.Collector.Models.WebSocket;
 
 
 namespace CCXT.Collector.Hitbtc