+++ src/exchanges/cn/okx/OkxWebSocketClient.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using CCXT.Collector.Core.Abstractions;
+using CCXT.Collector.Models.WebSocket;
 using CCXT.Collector.Library;
 using CCXT.Collector.Service;
 using System.Text.Json;