+++ src/exchanges/us/coinbase/CoinbaseWebSocketClient.cs
@@ -8,6 +8,7 @@ using CCXT.Collector.Core.Abstractions;
 using CCXT.Collector.Library;
 using CCXT.Collector.Service;
 using System.Text.Json;
+using CCXT.Collector.Models.WebSocket;
 
 namespace CCXT.Collector.Coinbase
 {